confirm delete logic updated

+++ app/lib/deleteAccount.ts
@@ -21,7 +21,7 @@ const crypto = require('crypto');
 import { sendMail } from './send-mail';
 import { dateCompare } from './dateCompare';
 const SITE_URL = process.env.SITE_URL;
-import { signOut } from '../../auth';
+//import { signOut } from '../../auth';
 import { customDeleteMailHtml, customDeleteMailText } from '../emails/deleteAccountEmail';
 
 const InitiateDeleteSchema = z.object({
@@ -60,11 +60,25 @@ export type State3 = {
 
 export async function autoSignOut(){
 
-  await signOut({redirectTo: '/account/delete/signed-out'})
+  //await signOut({redirectTo: '/account/delete/signed-out'})
+  return
 
 }
 
+import { headers } from 'next/headers';
+import { auth } from '../../auth';
+import { decryptUserData } from './encryption';
+
 export async function confirmDelete(prevState: State3, formData: FormData){
+  //retrieves session data for user
+  const session = await auth.api.getSession({
+    headers: await headers() // you need to pass the headers object.
+  })
+  
+  //redirects user to login if no session data
+  if (!session){    
+    return;
+  }
 
   //validates email and token to ensure not corrupted
   const validatedFields = NewConfirmUser.safeParse({    
@@ -89,16 +103,36 @@ export async function confirmDelete(prevState: State3, formData: FormData){
   const validatedEmail = validatedFields.data?.email;
   const validatedToken = validatedFields.data?.token;
   
+  //confirm email passed in token matches that for user signed in
+  const { email, id, encryptionDataId } = session.user;
+
+  const decryptedEmail = await decryptUserData(email, id, encryptionDataId);
+
+  /*This should be an auto signout situation, see previous logic */
+  if (decryptedEmail !== validatedEmail){
+    return {
+      message: 'Records do not match supplied credentials. You may need to send a new link.',
+      linkParams: 'wrong'
+    }
+  }
+
   //query and argument to confirm that the email and token are present and paired in the
   //database
   const confirmCredentialsQuery = 'SELECT * FROM delete_token WHERE identifier = $1 AND token = $2'
-  const confirmationArguments = [validatedEmail, validatedToken]
+  const confirmationArguments = [email, validatedToken]
 
   //queries for the deletion of all user data throughout the different tables
-  const deleteFromUsersQuery = 'DELETE FROM users WHERE email = $1 RETURNING *'
+  /*const deleteFromUsersQuery = 'DELETE FROM users WHERE email = $1 RETURNING *'
   const deleteFromSessionsQuery ='DELETE FROM sessions WHERE "userId" = $1' 
   const deleteFromVerificationTokenQuery = 'DELETE FROM verification_token WHERE identifier = $1'
+  const deleteFromDeleteTokenQuery = 'DELETE FROM delete_token WHERE identifier = $1'*/
+
+  const deleteFromUserQuery = 'DELETE FROM "user" WHERE id = $1 RETURNING *'
+  const deleteFromSessionQuery ='DELETE FROM session WHERE "userId" = $1'
+  //here identifier is currently the decrypted email, which is wrong
   const deleteFromDeleteTokenQuery = 'DELETE FROM delete_token WHERE identifier = $1'
+  const deleteFromAccountQuery = 'DELETE FROM account WHERE "userId" = $1'
+  const deleteFromEncryptionData = 'DELETE FROM encryption_data WHERE id = $1'
 
   try {
 
@@ -135,22 +169,32 @@ export async function confirmDelete(prevState: State3, formData: FormData){
       }
     }
 
-    //this deletes the user's entry in the users table, but also returns the
-    //corresponding id, which is then passed as userId (foreign key) to delete the user
-    //data from the sessions table
-    const usersReturned = await sql.query(deleteFromUsersQuery, [identifier])
+    
 
     //harvests userId foreign key (see above)
-    const userId = usersReturned.rows[0].id;
+    //const userId = usersReturned.rows[0].id;
 
     //deletes remaining user data from all other tables
+    /*
     await sql.query(deleteFromSessionsQuery, [userId])
     await sql.query(deleteFromVerificationTokenQuery, [identifier])
     await sql.query(deleteFromDeleteTokenQuery, [identifier])
+*/
+    await sql.query(deleteFromSessionQuery, [id])    
+    await sql.query(deleteFromDeleteTokenQuery, [email])
+    await sql.query(deleteFromAccountQuery, [id])
+    await sql.query(deleteFromEncryptionData, [encryptionDataId])
+
+    //this deletes the user's entry in the users table, but also returns the
+    //corresponding id, which is then passed as userId (foreign key) to delete the user
+    //data from the sessions table
+    //const usersReturned = await sql.query(deleteFromUsersQuery, [identifier])
+    await sql.query(deleteFromUserQuery, [id])
 
     await sql.query("COMMIT")    
 
   } catch (error){
+    console.log(error)
     
     await sql.query("ROLLBACK");
 
@@ -204,6 +248,15 @@ export async function checkExisting(email: string){
 }
 
 export async function initiateDelete(email: string, prevState: State2){
+  //retrieves session data for user
+  const session = await auth.api.getSession({
+    headers: await headers() // you need to pass the headers object.
+  })
+  
+  //redirects user to login if no session data (BEST URL FOR THIS?)
+  if (!session){    
+    redirect(`/account`);
+  }
   
   //validates email to ensure not corrupted, it expects one thing and one thing only
   //and that's a valid email
@@ -227,6 +280,8 @@ export async function initiateDelete(email: string, prevState: State2){
   //extracts validated email
   const validatedEmail = validatedFields.data?.email;
 
+  const {email: encryptedEmail} = session.user;
+
     //collectively, the two terms below generate a random 64-character alphanumerical string
     const token = crypto.randomBytes(32);
     const stringToken = token.toString('hex');    
@@ -239,7 +294,7 @@ export async function initiateDelete(email: string, prevState: State2){
        ON CONFLICT (identifier, token) DO UPDATE 
          SET expires = excluded.expires, 
          token = excluded.token;`
-    const argumentData = [validatedEmail, timestamp, stringToken];
+    const argumentData = [encryptedEmail, timestamp, stringToken];
 
     try {
